Move patient registration components into register folder
