+++ client/src/pages/patients/register/PatientRegistationAccordion.module.css
