+++ client/src/pages/patients/register/PatientRegistrationAccordion.jsx
