+++ client/src/pages/patients/register/inputs/SearchDatabaseInputField.jsx
